chore: code refractored
