+++ packages/suite/src/config/suite/settings.ts
@@ -1,3 +1,16 @@
+import { NotificationEntry } from '@suite-reducers/notificationReducer';
+
+const IMPORTANT_NOTIFICATION_TYPES: Array<NotificationEntry['type']> = [
+    'tx-sent',
+    'tx-received',
+    'tx-confirmed',
+    'clear-storage',
+    'pin-changed',
+    'device-wiped',
+    'backup-success',
+    'backup-failed',
+];
+
 export default {
     MAX_ACCOUNTS: 10,
     FRESH_ADDRESS_LIMIT: 20,
@@ -8,4 +21,5 @@ export default {
         endDate: null,
         groupBy: 'month',
     },
+    IMPORTANT_NOTIFICATION_TYPES,
 } as const;
