+++ packages/suite/src/utils/suite/notification.ts
@@ -1,5 +1,5 @@
 import { NotificationEntry } from '@suite-reducers/notificationReducer';
-import { ToastNotificationVariant } from '@suite-types';
+import { ToastNotificationVariant, AppState } from '@suite-types';
 
 export const findTransactionEvents = (descriptor: string, notifications: NotificationEntry[]) =>
     notifications.filter(
@@ -20,4 +20,20 @@ export const getNotificationIcon = (variant: ToastNotificationVariant) => {
             return 'CHECK';
         // no default
     }
+};
+
+export const getSeenAndUnseenNotifications = (notifications: AppState['notifications']) => {
+    const seen: Array<NotificationEntry> = [];
+    const unseen: Array<NotificationEntry> = [];
+
+    // loop over all notifications and check which of them there were seen or not
+    notifications.forEach(notification => {
+        if (notification.seen) {
+            seen.push(notification);
+        } else {
+            unseen.push(notification);
+        }
+    });
+
+    return { seenNotifications: seen, unseenNotifications: unseen };
 };
